Allow null FallID in Labordaten
The provided test data in SmICSCore returns null for some of the results
for the Labordaten query, which is not allowed in the JSON schema. This
leads to a rejection of the data returned from SmICSCore, so that the
lab data is missing in the cache. This breaks modules dependent on that
data.